% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{generate_embeddings}
\alias{generate_embeddings}
\title{Generate Embeddings}
\usage{
generate_embeddings(
  session,
  inputs,
  output_index = 0,
  mean_pooling_needed = FALSE
)
}
\arguments{
\item{session}{an Inference Session object}

\item{inputs}{a vector of string inputs you want to calculate the embeddings for}
}
\description{
Generate Embeddings
}
