% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{inference_session_from_file}
\alias{inference_session_from_file}
\title{Initialize a new Inference Session from a saved model}
\usage{
inference_session_from_file(model_path, tokenizer_path, num_threads = 1)
}
\arguments{
\item{model_path}{the path to where the onnx model is stored}

\item{tokenizer_path}{the path to the tokenizer is stored}

\item{num_thread}{number of threads for inference to run on (default 1)}
}
\description{
Initialize a new Inference Session from a saved model
}
