% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{new_inference_session}
\alias{new_inference_session}
\title{Initialize a new Inference Session from a saved model}
\usage{
new_inference_session(
  directory,
  model_name = "model.onnx",
  tokenizer_name = "tokenizer.json"
)
}
\arguments{
\item{model_path}{the path to where the onnx model is stored}

\item{tokenizer_path}{the path to the tokenizer is stored}

\item{num_thread}{number of threads for inference to run on (default 1)}
}
\description{
Initialize a new Inference Session from a saved model
}
