% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_models.R
\name{race_models}
\alias{race_models}
\title{Compare performance between different models}
\usage{
race_models(model_list, n = 1000)
}
\arguments{
\item{model_list}{a named list of different inference sessions you would
like to compare}

\item{n}{the size of the dataset used for evaluation.}
}
\description{
Compare performance between different models
}
