% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{run_model}
\alias{run_model}
\title{Run a Model on Given Inputs}
\usage{
run_model(session, inputs)
}
\arguments{
\item{session}{an Inference Session object}

\item{inputs}{a vector of string inputs you want to calculate the embeddings for}
}
\description{
Run a Model on Given Inputs
}
